% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_impd_fhx.R
\name{get_impd_fhx}
\alias{get_impd_fhx}
\title{Retrieve tree-ring fire history data from the IMPD}
\usage{
get_impd_fhx(studyCode)
}
\arguments{
\item{studyCode}{Provides a unique reference to retrieve an FHX file from the
IMPD. The \code{studyCode} field is available in the data frame generated by \code{\link[=build_impd_meta]{build_impd_meta()}}}
}
\description{
Retrieve tree-ring fire history data from the IMPD
}
\examples{
fsr <- get_impd_fhx("USFSR001")
burnr::plot(fsr)
}
