% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impd_meta}
\alias{impd_meta}
\title{IMPD search results}
\format{
A data frame with 15 variables: \code{siteName}, \code{studyCode},
\code{investigators}, \code{first_year}, \code{last_year}, \code{latitude},\code{longitude},
\code{elevation}, \code{species}, \code{reference}, \code{doi}, \code{contr_year},
\code{NOAAStudyID}, \code{NOAASiteID}, and \code{url}.
}
\usage{
impd_meta
}
\description{
Full list of North American tree-ring records in the IMPD
}
\keyword{datasets}
