% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncei_paleo_api.R
\name{ncei_paleo_api}
\alias{ncei_paleo_api}
\title{IMPD records search}
\usage{
ncei_paleo_api(
  investigators = NULL,
  minLat = NULL,
  maxLat = NULL,
  minLon = NULL,
  maxLon = NULL,
  minElev = NULL,
  maxElev = NULL,
  earliestYear = NULL,
  latestYear = NULL,
  species = NULL
)
}
\arguments{
\item{investigators}{Name of investigator listed by the IMPD. The format can
look like "Swetnam", "Swetnam, T", or "Swetnam, T.W.".  See
\code{\link[=get_search_params]{get_search_params()}} for a list potential names}

\item{minLat}{Minimum latitude of sites, in decimal degrees (WGS84)}

\item{maxLat}{Maximum latitude of sites, in decimal degrees (WGS84)}

\item{minLon}{Minimum longitude of sites, in decimal degrees (WGS84)}

\item{maxLon}{Maximum longitude of sites, in decimal degrees (WGS84)}

\item{minElev}{Minimum elevation of sites, in meters}

\item{maxElev}{Maximum elevation of sites, in meters}

\item{earliestYear}{Lower dated bound of sites. Use calendar years CE (negative for BCE)}

\item{latestYear}{Upper dated bound of sites. Use calendar years CE (negative for BCE)}

\item{species}{Four-letter species code for tree species. See
\code{\link[=get_search_params]{get_search_params()}} to find the proper code}
}
\description{
Query the IMPD for North American tree-ring fire-scar records
}
\examples{
# Search the IMPD for Tom Swetnam's contributions
swetnam_search <- ncei_paleo_api(investigators = "Swetnam")
}
\seealso{
\code{\link[=get_search_params]{get_search_params()}}
}
