% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quga_dat}
\alias{quga_dat}
\title{IMPD example data}
\format{
A data frame with 16 variables: \code{siteName}, \code{studyCode},
\code{investigators}, \code{first_year}, \code{last_year}, \code{latitude},\code{longitude},
\code{elevation}, \code{species}, \code{reference}, \code{doi}, \code{contr_year},
\code{NOAAStudyID}, \code{NOAASiteID}, \code{url}, and \code{FHX}.
}
\usage{
quga_dat
}
\description{
IMPD records with QUGA (Gambel oak) as a species, including FHX files
}
\keyword{datasets}
