% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_impd.R
\name{search_impd}
\alias{search_impd}
\title{Search the IMPD}
\usage{
search_impd(
  investigators = NULL,
  minLat = NULL,
  maxLat = NULL,
  minLon = NULL,
  maxLon = NULL,
  minElev = NULL,
  maxElev = NULL,
  earliestYear = NULL,
  latestYear = NULL,
  species = NULL
)
}
\arguments{
\item{investigators}{Name of investigator listed by the IMPD. The format can
look like "Swetnam", "Swetnam, T", or "Swetnam, T.W.".  See
\link{get_search_params("investigators")} for a list potential names}

\item{minLat}{Minimum latitude of sites, in decimal degrees (WGS84)}

\item{maxLat}{Maximum latitude of sites, in decimal degrees (WGS84)}

\item{minLon}{Minimum longitude of sites, in decimal degrees (WGS84)}

\item{maxLon}{Maximum longitude of sites, in decimal degrees (WGS84)}

\item{minElev}{Minimum elevation of sites, in meters}

\item{maxElev}{Maximum elevation of sites, in meters}

\item{earliestYear}{In calendar years CE (use negative for BCE)}

\item{latestYear}{In calendar years CE (use negative for BCE)}

\item{species}{Four-letter species code for tree species. See
\link{get_search_params("species")} for a data frame of Latin names and codes}
}
\value{
A data frame of search results form the North American IMPD
}
\description{
This function provides a means for fast and efficient searches of the North
American tree-ring fire scar datasets housed on the International Multiproxy
Paleofire Database (IMPD)
}
\examples{
# Find Tom Swetnam's fire-history sites
search_impd(investigators = "Swetnam")

# Find all sites with Gambel oak
search_impd(species = "QUGA")
}
