% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ncei_paleo_api.R
\name{search_ncei_paleo_api}
\alias{search_ncei_paleo_api}
\title{Search the NCEI Paleo API for IMPD records}
\usage{
search_ncei_paleo_api(
  investigators = NULL,
  minLat = NULL,
  maxLat = NULL,
  minLon = NULL,
  maxLon = NULL,
  minElev = NULL,
  maxElev = NULL,
  earliestYear = NULL,
  latestYear = NULL,
  species = NULL
)
}
\arguments{
\item{investigators}{Name of investigator listed by the IMPD}

\item{minLat}{Minimum latitude of sites, in decimal degrees (WGS84)}

\item{maxLat}{Maximum latitude of sites, in decimal degrees (WGS84)}

\item{minLon}{Minimum longitude of sites, in decimal degrees (WGS84)}

\item{maxLon}{Maximum longitude of sites, in decimal degrees (WGS84)}

\item{minElev}{Minimum elevation of sites, in meters}

\item{maxElev}{Maximum elevation of sites, in meters}

\item{earliestYear}{In calendar years CE (use negative for BCE)}

\item{latestYear}{In calendar years CE (use negative for BCE)}

\item{species}{Use four-letter codes for tree species}
}
\description{
Query the IMPD for fire-scar records based on search parameters
}
\examples{
# Search the IMPD for Tom Swetnam's contributions
swetnam_search <- search_ncei_paleo_api(investigators = "Swetnam")
}
