% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_impd_meta.R
\name{build_impd_meta}
\alias{build_impd_meta}
\title{Assemble tidy table from IMPD search results}
\usage{
build_impd_meta(api)
}
\arguments{
\item{api}{Resulting object from \code{\link[=search_ncei_paleo_api]{search_ncei_paleo_api()}}}
}
\description{
This function extracts desired information form the search results to the
IMPD and returns a tidy data frame of the results
}
\examples{
api <- search_ncei_paleo_api("Swetnam")
meta <- build_impd_meta(api)
}
