% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_impd_fhx.R
\name{get_impd_fhx}
\alias{get_impd_fhx}
\title{Retrieve tree-ring fire history data from the IMPD}
\usage{
get_impd_fhx(studyCode)
}
\arguments{
\item{studyCode}{Provides a unique reference to retrieve an FHX file from the
IMPD. The \code{studyCode} field is available in the data frame generated by
\code{\link[=build_impd_meta]{build_impd_meta()}}}
}
\value{
Retrieved files are in FHX format, which are read in and formatted
with the \code{burnr} library. See
\url{https://ltrr-arizona-edu.github.io/burnr/reference/read_fhx.html}
for more details.
}
\description{
Retrieve tree-ring fire history data from the IMPD
}
\details{
The function pulls FHX file(s) from the NOAA ftp server at
\url{https://www.ncei.noaa.gov/pub/data/paleo/firehistory/firescar/northamerica/}.
Therefore, the success of the function largely depends on the availability
of the server. Sometimes it may not work due to network issues, so patience
may be required. Currently, the function limits searches to a single study
code. Use loops, or \code{purrr}-style iterative calls to expedite the retrieval
of multiple files.
}
\examples{
fsr <- get_impd_fhx("USFSR001")
burnr::plot(fsr)
}
