% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_params.R
\name{get_search_params}
\alias{get_search_params}
\title{Extract IMPD search parameters for \code{investigators} and \code{species}}
\usage{
get_search_params(output = c("investigators", "species"))
}
\arguments{
\item{output}{Two data sets are available, "investigators" (the default) and "species"}
}
\description{
This function pulls form the updated store of contributor names and species
to help users specify their searches
}
\examples{
# Retrieve the list of IMPD contributors
get_search_params("investigators")

# Retrieve a data frame of species codes
get_search_params("species")
}
