% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_params.R
\name{get_search_params}
\alias{get_search_params}
\title{Extract IMPD search parameters for \code{investigators}, \code{species}, and \code{location}}
\usage{
get_search_params(output)
}
\arguments{
\item{output}{Three data sets are available: "investigators" (the default),
"location", and "species"}
}
\description{
Retrieve the updated store of contributor names, site locations, and tree
species to help specify searches of the International Multiproxy
Paleofire Database (IMPD).
}
\note{
The list of investigators names also includes those who have
contributed to the charcoal and pollen databases of the IMPD. Some names
may not generate search results here because \code{rIMPD} specifies tree-ring
records. Also, apologies to Canadians for using "state" as a
variable name in the output for "location"
}
\examples{
# Retrieve the list of IMPD contributors
get_search_params("investigators")

# Retrieve a data frame of species codes
get_search_params("species")
}
