% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_impd.R
\name{search_impd}
\alias{search_impd}
\title{Search the IMPD}
\usage{
search_impd(
  investigators = NULL,
  location = NULL,
  minLat = NULL,
  maxLat = NULL,
  minLon = NULL,
  maxLon = NULL,
  minElev = NULL,
  maxElev = NULL,
  earliestYear = NULL,
  latestYear = NULL,
  species = NULL
)
}
\arguments{
\item{investigators}{Name of investigator listed by the IMPD. The format can
look like "Swetnam", "Swetnam, T", or "Swetnam, T.W.".  See
\code{\link[=get_search_params]{get_search_params()}} for a list potential names}

\item{location}{The country or state/province of sites. Entries require
proper spelling, for example, use "Quebec" or "United States Of America"
and not "arizona". Not all localities are available; check
\code{\link[=get_search_params]{get_search_params()}} for availability of your place of interest.}

\item{minLat}{Minimum latitude of sites, in decimal degrees (WGS84)}

\item{maxLat}{Maximum latitude of sites, in decimal degrees (WGS84)}

\item{minLon}{Minimum longitude of sites, in decimal degrees (WGS84)}

\item{maxLon}{Maximum longitude of sites, in decimal degrees (WGS84)}

\item{minElev}{Minimum elevation of sites, in meters}

\item{maxElev}{Maximum elevation of sites, in meters}

\item{earliestYear}{Lower dated bound of sites. Use calendar years CE
(negative for BCE)}

\item{latestYear}{Upper dated bound of sites. Use calendar years CE (negative
for BCE)}

\item{species}{Four-letter species code for tree species. See
\code{\link[=get_search_params]{get_search_params()}} to find the proper code}
}
\value{
A data frame of search results from the North American IMPD,
including: \itemize{
\item{"siteName" -- name of the site}
\item{"studyCode" -- unique site identifier as defined by the IMPD}
\item{"investigators" -- list of investigators for the study site}
\item{"first_year" -- earliest year of the tree-ring data}
\item{"last_year" -- latest year of the tree-ring data}
\item{"latitude" and "longitude"  -- site coordinates}
\item{"elevation" -- site elevation in meters above sea level}
\item{"species" -- tree species codes}
\item{"reference" -- literature citation for site data}
\item{"doi" -- Digital Object Identifier for the site data}
\item{"contr_year" -- the year the site data were contriuted to the IMPD}
\item{"NOAAStudyId" -- unique study identifier used by NOAA}
\item{"NOAASiteId" -- unique site identifier used by NOAA}
\item{"url" -- site landing page on the NCEI server}
}
}
\description{
This function provides a means for fast and efficient searches of the North
American tree-ring fire scar datasets housed on the International Multiproxy
Paleofire Database (IMPD)
}
\note{
If you search the IMPD a lot, you might find inconsistent results. For
one, the IMPD is constantly being updated with new sites and thus growing.
But sometimes not all sites will appear in a search that had been showing
up previously. This is not a issue in \code{rIMPD}, rather it has to do with the
NCEI API being accessed by \code{rIMPD}. Should you have trouble with the
search, or notice data errors in the results, then email "paleo at
noaa.gov". The WDS-Paleo team has been very helpful in building this
package and are always working to improve the API and paleo resources.
}
\examples{
# Find Tom Swetnam's fire-history sites
search_impd(investigators = "Swetnam")

# Find all sites with Gambel oak
search_impd(species = "QUGA")

# Retrieve sites from Oregon and Washington states
or_wa <- rbind(search_impd(location = "Oregon"),
               search_impd(location = "Washington"))
}
